Authentication

+++ backend/routes/posts.js
@@ -65,7 +65,8 @@ router.post("",
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then(createdPost => {
     res.status(201).json({
@@ -90,18 +91,25 @@ router.patch("/:id",
     _id: req.body.id,
     title: req.body.title, 
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
-  Post.updateOne({ _id: req.params.id}, post).then(result => {
-    console.log(result)
-    res.status(200).json({ message: "Update Successfully" })
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+    if (result.nModified > 0) {
+      res.status(200).json({ message: "Update Successfully" })
+    } else {
+      res.status(401).json({ message: "Not Authorized!" })
+    }
   })
 })
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id}).then(result => {
-    console.log(result);
-    res.status(200).json({ message: "Post deleted successfully"});
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+    if (result.n > 0) {
+      res.status(200).json({ message: "Delete Successfully" })
+    } else {
+      res.status(401).json({ message: "Not Authorized!" })
+    }
   })
 });
